test: adding extend tests, tweaking
