+++ lib/index.js
@@ -3,6 +3,7 @@ const webpackLog = require('webpack-log');
 const validate = require('@webpack-contrib/schema-utils');
 const webpackSchema = require('webpack/schemas/WebpackOptions.json');
 
+const { extend } = require('./extend');
 const load = require('./load');
 const resolve = require('./resolve');
 
@@ -13,11 +14,14 @@ module.exports = (options = {}) => {
   const raw = load(options);
 
   return resolve(raw).then((result) => {
-    const schema = merge({}, options.schema, webpackSchema);
-    for (const target of [].concat(result.config)) {
-      validate({ name, schema, target });
-    }
+    const { config, configPath } = result;
 
-    return result;
+    return extend(config).then((finalConfig) => {
+      const schema = merge({}, options.schema, webpackSchema);
+      for (const target of [].concat(finalConfig)) {
+        validate({ name, schema, target });
+      }
+      return { config: finalConfig, configPath };
+    });
   });
 };
