+++ package.json
@@ -82,7 +82,7 @@
     "pre-commit": "^1.2.2",
     "prettier": "^1.11.1",
     "standard-version": "^4.3.0",
-    "ts-node": "^7.0.0",
+    "ts-node": "^5.0.1",
     "typescript": "^2.8.1",
     "webpack": "^4.4.1",
     "webpack-defaults": "^2.3.0"
