+++ test/test.js
@@ -8,3 +8,4 @@ require('./tests/load');
 require('./tests/resolve');
 require('./tests/full');
 require('./tests/errors');
+require('./tests/extends');
